// src/app/pages/Model/CustomerMaster.Model.ts
export class CustomerMasters {
    Id: any;
    App_Code:any;
    Customer_ID: any;
    Customer_Name: any;
    Email: any;
    Mobile: any;
    PermanentAddress: any;
    Pincode: any;  
    landmark : any;     
    UserName: any;
    Password: any;
    Posted_Date: Date = new Date();
    Notes: any;
    IsDeactive: any = 1;      
}    